Move process_concatenated_datasets to helpers file
