+++ src/deepsparse/transformers/eval_downstream.py
@@ -67,10 +67,10 @@ from pstats import Stats
 
 import numpy
 from tqdm.auto import tqdm
-from transformers import AutoTokenizer
 
 from deepsparse import DEEPSPARSE_ENGINE, ORT_ENGINE, Pipeline
 from deepsparse.transformers.metrics import Perplexity, PrecisionRecallF1
+from deepsparse.transformers.utils.eval_helpers import process_concatenated_datasets
 
 
 from datasets import load_dataset, load_metric  # isort: skip
@@ -83,7 +83,7 @@ def perplexity_eval(args, dataset_name="openai_humaneval"):
             kwargs = {}
         else:
             kwargs = json.loads(args.kwargs)
-        dataset = _process_concatenated_datasets(
+        dataset = process_concatenated_datasets(
             dataset_name,
             args.model_path,
             args.max_sequence_length,
@@ -520,77 +520,6 @@ def _split_train_val(train_dataset, val_ratio, seed=42):
     train_ds = ds.pop("train")
     val_ds = ds.pop("test")
     return train_ds, val_ds
-
-
-def _process_concatenated_datasets(dataset_name, model_path, max_sequence_length, kwargs):
-    if dataset_name == "wikitext2":
-        eos = kwargs.get("eos", "\n\n")
-        bos = kwargs.get("bos", "")
-
-        raw_dataset = load_dataset("wikitext", "wikitext-2-raw-v1", split="test")
-        raw_text = raw_dataset["text"]
-    elif dataset_name == "c4":
-        eos = kwargs.get("eos", "<|endoftext|>")
-        bos = kwargs.get("bos", "")
-        raw_samples = kwargs.get("raw_samples", None)
-        data_file = kwargs.get("data_file", 0)
-        if data_file is not None:
-            raw_dataset = load_dataset(
-                "allenai/c4",
-                "allenai--c4",
-                data_files={"validation": f"en/c4-validation.{data_file:05d}-of-00008.json.gz"},
-                split="validation",
-            )
-        else:
-            raw_dataset = load_dataset(
-                "allenai/c4",
-                "allenai--c4",
-                split="validation",
-            )
-        if raw_samples is not None:
-            raw_dataset = raw_dataset[:raw_samples]
-        raw_text = raw_dataset["text"]
-
-    # Dataset is split into sections that contain "max_sequence_length" tokens.
-    # To split the dataset, first tokenize text
-    tokenizer = AutoTokenizer.from_pretrained(model_path)
-    return _split_text_by_tokens(
-        raw_text, eos, bos, tokenizer, max_sequence_length,
-    )
-
-
-def _split_text_by_tokens(text, eos, bos, tokenizer, sequence_length):
-    text = "".join([bos + sample + eos for sample in text])
-
-    input_tokens = tokenizer(text, return_tensors="np",)[
-        "input_ids"
-    ][0]
-
-    # Then split the tokenized text into sections of size "max_sequence_length" and
-    # decode each section back into text format
-    split_text = []
-    for i in range(len(input_tokens) // sequence_length):
-        start = i * sequence_length
-        end = (i + 1) * sequence_length
-        split_text.append(
-            tokenizer.decode(
-                input_tokens[start:end],
-                clean_up_tokenization_spaces=False,
-            )
-        )
-
-    # Handle any leftover tokens
-    if (i + 1) * sequence_length < len(input_tokens):
-        start = (i + 1) * sequence_length
-        end = len(input_tokens)
-        split_text.append(
-            tokenizer.decode(
-                input_tokens[start:end],
-                clean_up_tokenization_spaces=False,
-            )
-        )
-
-    return split_text
 
 
 # Register all the supported downstream datasets here
